add signal handler for POSIX signal SIGHUP to force a hard stumpwm restart
FORCE-STUMPWM-RESTART will immediately force a restart outside stumpwm's
loops and is kind of harsh.

Sometimes when debugging I cannot invoke the command RESTART-HARD anymore.
But you cannot call RESTART-HARD from a swank as it needs to be in the
context of the CATCH :TOP-LEVEL.

The :HUP-PROCESS in function STUMPWM gave me the idea that stumpwm should
even do this when receiving SIGHUP and this is kind of nice so that you
can restart stumpwm even when you cannot invoke commands properly and you
do not have a running swank server anymore (which sometimes happens for me).